requirements.txt
amqp==5.3.1
anyio==4.10.0
asgiref==3.8.1
billiard==4.2.1
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==45.0.6
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.6
django-celery-beat==2.8.1
django-cors-headers==4.4.0
django-heroku==0.3.1
django-mssql==1.8
django-mssql-backend==2.8.1
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
firebase_admin==7.1.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==22.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
kombu==5.5.4
msgpack==1.1.1
mssql-django==1.5
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.32.0
psycopg2==2.9.9
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyodbc==5.1.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==6.2.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
virtualenv==20.26.3
virtualenvwrapper-win==1.2.7
wcwidth==0.2.13
whitenoise==6.7.0
